Add local ids for pasted fields
