+++ src/utilities/form-helpers/add-local-ids.ts
@@ -17,6 +17,7 @@ const addLocalIds = {
                 },
             );
         }
+        return field;
     },
 };
