+++ src/utilities/template-helpers/add-field.ts
@@ -1,15 +1,4 @@
-import {configTypes} from 'utilities';
-import {max} from 'lodash';
-
-function newFieldId(config: configTypes.SurveyConfig): number {
-    const maxId = max(
-        config.fields.map((field: configTypes.SurveyField) => field.local_id),
-    );
-    if (maxId === undefined) {
-        return config.local_id;
-    }
-    return maxId + 1;
-}
+import {configTypes, newFieldId} from 'utilities';
 
 const fieldTemplate = (
     fieldType: configTypes.FieldType,
